+++ CHANGELOG.md
@@ -18,8 +18,9 @@ Note that this project **does not** adhere to [Semantic Versioning](http://semve
 
 ### Changed
 
+- 'Get full text' now also checks the file url. [#568](https://github.com/koppor/jabref/issues/568)
 - We refined the 'main directory not found' error message. [#9625](https://github.com/JabRef/jabref/pull/9625)
-
+- We modified the `Add Group` dialog to use the most recently selected group hierarchical context [#9141](https://github.com/JabRef/jabref/issues/9141)